🏗️ UsePostQueries 위치 이동
