+++ src/entities/post/api/hooks/usePostQueries.ts
@@ -1,5 +1,5 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
-import { postApi, PostsTypes } from "../postApi"
+import { useQuery } from "@tanstack/react-query"
+import { postApi } from "../postApi"
 import { postQueryKeys } from "../postQueryKeys"
 
 export const useGetPosts = (limit: number, skip: number) => {
@@ -21,53 +21,5 @@ export const useGetPostsByTag = (tag: string) => {
     queryKey: postQueryKeys.tags(tag),
     queryFn: () => postApi.getPostsByTag(tag),
     enabled: !!tag,
-  })
-}
-
-export const useUpdatePost = () => {
-  const queryClient = useQueryClient()
-  const queryKey = postQueryKeys.lists()
-
-  return useMutation({
-    mutationFn: postApi.updatePost,
-    onMutate: async (updatedPost) => {
-      const previousPosts = queryClient.getQueryData(queryKey)
-
-      queryClient.setQueryData(queryKey, (old: PostsTypes["posts"] = []) => {
-        return old.map((post) => (post.id === updatedPost.id ? updatedPost : post))
-      })
-
-      return { previousPosts }
-    },
-    onError: (err, _, context) => {
-      queryClient.setQueryData(queryKey, context?.previousPosts)
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: queryKey })
-    },
-  })
-}
-
-export const useDeletePost = () => {
-  const queryClient = useQueryClient()
-  const queryKey = postQueryKeys.lists()
-
-  return useMutation({
-    mutationFn: postApi.deletePost,
-    onMutate: async (postId) => {
-      const previousPosts = queryClient.getQueryData(queryKey)
-
-      queryClient.setQueryData(queryKey, (old: PostsTypes["posts"] = []) => {
-        return old.filter((post) => post.id !== postId)
-      })
-
-      return { previousPosts }
-    },
-    onError: (err, _, context) => {
-      queryClient.setQueryData(queryKey, context?.previousPosts)
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: queryKey })
-    },
   })
 }
